Refresh complete
Re-implemented this project using newly refreshed MongoCRUD as a
submodule.  No new functionality, but is now much nicer to install and
setup (8 years of experience will do that).  Can now be run in dev mode,
as a container, or a compose app, and can generate a dist zip.

There are still lots of improvements that need to be made and /hot is
not really done, but I've gotten to parity with the old version.